// app/app/page.module.css.d.ts
declare const styles: {
  container: string;
  header: string;
  headerContent: string;
  logo: string;
  logoText: string;
  headerRight: string;
  statusBadge: string;
  settingsLink: string;
  srOnly: string;
  main: string;
  grid: string;
  sendSection: string;
  connectionSection: string;
  sendCard: string;
  connectionCard: string;
  devicesCard: string;
  cardTitle: string;
  cardDescription: string;
  dropZone: string;
  dropZoneActive: string;
  dropZoneWithFiles: string;
  dropZoneContent: string;
  dropZoneText: string;
  dropZoneHint: string;
  fileInput: string;
  filesList: string;
  filesHeader: string;
  filesCount: string;
  filesItems: string;
  fileItem: string;
  fileInfo: string;
  fileName: string;
  fileSize: string;
  removeButton: string;
  codeSection: string;
  codeLabel: string;
  codeDisplay: string;
  codeText: string;
  copyButton: string;
  qrCode: string;
  qrCodePlaceholder: string;
  qrCodeText: string;
  divider: string;
  dividerText: string;
  connectSection: string;
  connectedDevice: string;
  deviceAvatar: string;
  deviceInitial: string;
  deviceInfo: string;
  deviceName: string;
  devicePlatform: string;
  securityIndicators: string;
  securityItem: string;
  securityText: string;
  devicesList: string;
  deviceButton: string;
  transferProgress: string;
  progressHeader: string;
  progressInfo: string;
  progressTitle: string;
  progressStats: string;
  progressDot: string;
  progressBarContainer: string;
  progressBar: string;
  progressPercentage: string;
};

export default styles;